mcp/zotero/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paperpowers-zotero-mcp"
version = "0.1.0"
description = "A lightweight MCP server for Zotero used by PaperPowers"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "mcp[cli]",
  "python-dotenv",
  "pyzotero",
]
